add warmup_rounds parameter, kill existing workers on setup
When there were old workers lingering, the results would not be correct,
as there would be more worker processes running per node than available cores.

The warmup_rounds parameter is an attempt to improve performance of
in-memory mmapped operations, as it can take a few rounds to make sure
that all data is cached in fs cache.